requirements.txt
streamlit==1.32.0
pandas==2.2.0
openai==1.12.0
python-dotenv==1.0.1
plotly==5.18.0
numpy==1.26.4
python-magic==0.4.27
openpyxl==3.1.2
tenacity==8.2.3 